requirements.txt
# Standard
numpy
matplotlib
seaborn
pandas

# Code quality
mypy
pytest
flake8
brunette

# Learning libraries
#torch==1.9.0+cu111
# torch==1.9.0
pytorch-lightning==1.3.4
torchmetrics==0.5.1

# Optimization
cvxpy==1.2.1
cvxpylayers==0.1.5
gurobipy==9.1.2
gurobipy
osqp
ecos
scs
casadi

# Geometry
shapely

# Freeze versions to avoid bugs
setuptools==59.5.0

# added for __div__ error
scikit-learn==1.2.1
scipy==1.10.1
pillow==9.5.0
